Alter structure, solve issues with some graphs
Problem:
- Memory that is not freed (not a leak and not detected in Valgrind)
- Problem with a type of graph where we might have an encapsulation of
  component $C1$ of an other component $C2$ with vertex $v$ and that
  not all of its vertices have a path to $l(v)-1$

Solution:
- Change types for better message complexity.
- Test only one graph per execution and add a shell script for making
  all the executions needed for our tests.
- Alter CMAKEFILE to copy the shell script to the build directory.
- Implement Hanani-Tutte differently to take in consideration the
  problem with the encapsulation.
- CMAKEFILE add a shell script for the testes and copy the shell script

Impact:
- Less space consumption with flat_map from boost but still not enough
  to make to execute all the tests in one execution, we still have the
issue of the freed memory not being sent back to the os.
- Using a shell script and restricting one graph per execution made sure
  that the freed memory each time get sent back to the OS. Now the issue
is with a single graph execution, to some point we can't create bigger
graphs. (more specifically if we have big number of nodes and smaller
number of levels).
- The new Implementation of hanani-tutte fixed the issue, but it is not
  a correct solution as it made clear that there is an other case that
is not taking in consideration by it, which is in case we have a vertex
$v \in C_1$ and $v$ has no path to the minimum level of $C_2$. In that
case and if we have a vertex $w \in C_2$ with no neighbhood in $l(v)$
and $l(v) > l(w)$, the new 'propagation' doesn't propagate to the
relation $\left(v,w\right)$.